scplan: move stage logic to new scstage package
This commit is strictly a refactor which moves all stage planning logic
from the declarative schema changer into its own package, with no change
to the structure of the code or the logic in the function bodies.

Release note: None
