+++ pkg/sql/schemachanger/scplan/scstage/job_augmentation.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package scplan
+package scstage
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/jobs"
@@ -22,7 +22,7 @@ import (
 	"github.com/cockroachdb/errors"
 )
 
-// augmentedStagesForJob adds operations to the stages to deal with updating
+// AugmentStagesForJob adds operations to the stages to deal with updating
 // the job status and for updating the references from descriptors to jobs.
 //
 // TODO(ajwerner): Rather than adding this above the opgen layer, it'd be
@@ -32,7 +32,9 @@ import (
 // may prove to be a somewhat common pattern in other cases: consider the
 // intermediate index needed when adding and dropping columns as part of the
 // same transaction.
-func augmentStagesForJob(params Params, stages []Stage) (augmented []Stage, jobID jobspb.JobID) {
+func AugmentStagesForJob(
+	stages []Stage, executionPhase scop.Phase, jobIDGenerator func() jobspb.JobID,
+) (augmented []Stage, jobID jobspb.JobID) {
 	var (
 		toProcess   = stages
 		addStage    = func(s Stage) { augmented = append(augmented, s) }
@@ -49,7 +51,7 @@ func augmentStagesForJob(params Params, stages []Stage) (augmented []Stage, jobI
 
 		descIDs []descpb.ID
 		setup   = func() (done bool) {
-			switch params.ExecutionPhase {
+			switch executionPhase {
 			case scop.StatementPhase:
 				augmented = toProcess
 				return true // done
@@ -73,18 +75,18 @@ func augmentStagesForJob(params Params, stages []Stage) (augmented []Stage, jobI
 				}
 				var opsToAdd []scop.Op
 				jobID, descIDs, opsToAdd = createSchemaChangeJobAndAddDescriptorJobReferenceOps(
-					params.JobIDGenerator, toProcess[0].After, toProcess[0].Revertible,
+					jobIDGenerator, toProcess[0].After, toProcess[0].Revertible,
 				)
 				addCurStage(opsToAdd...)
 				return false // done
 			case scop.PostCommitPhase:
 				// Initialize the jobID and descriptors IDs to update for the
 				// PostCommit stages.
-				jobID, descIDs = params.JobIDGenerator(), descIDsFromState(toProcess[0].Before)
+				jobID, descIDs = jobIDGenerator(), descIDsFromState(toProcess[0].Before)
 				return false // done
 
 			default:
-				panic(errors.AssertionFailedf("unknown phase %v", params.ExecutionPhase))
+				panic(errors.AssertionFailedf("unknown phase %v", executionPhase))
 			}
 		}
